+++ core/src/cli/orakl-cli/src/cli-types.ts
@@ -25,3 +25,45 @@ export const ReadFile: Type<string, string> = {
     return JSON.parse((await loadFile(filePath)).toString())
   }
 }
+
+interface IHeader {
+  'Content-Type': string
+}
+
+interface IReducer {
+  function: string
+  args: string[]
+}
+
+interface IFeed {
+  url: string
+  headers?: IHeader[]
+  method: string
+  reducers?: IReducer[]
+}
+
+export interface IAdapter {
+  id?: string
+  active?: boolean
+  name?: string
+  jobType?: string
+  decimals: number
+  feeds: IFeed[]
+}
+
+interface IProperty {
+  active: boolean
+  value: number
+}
+
+export interface IAggregator {
+  id: string
+  active: boolean
+  name: string
+  aggregatorAddress: string
+  fixedHeartbeatRate: IProperty
+  randomHeartbeatRate: IProperty
+  threshold: number
+  absoluteThreshold: number
+  adapterId: string
+}
