+++ core/src/worker/utils.ts
@@ -1,21 +1,11 @@
 import { ethers } from 'ethers'
 import { Logger } from 'pino'
 import { OraklError, OraklErrorCode } from '../errors'
-import { IAdapter, IAggregator, IOracleRoundState, IRoundData } from '../types'
+import { IOracleRoundState, IRoundData } from '../types'
 import { PROVIDER } from '../settings'
 import { Aggregator__factory } from '@bisonai/orakl-contracts'
 
 const FILE_NAME = import.meta.url
-
-function checkDataFormat(data) {
-  if (!data) {
-    // check if priceFeed is null, undefined, NaN, "", 0, false
-    throw new OraklError(OraklErrorCode.InvalidDataFeed)
-  } else if (!Number.isInteger(data)) {
-    // check if priceFeed is not Integer
-    throw new OraklError(OraklErrorCode.InvalidDataFeedFormat)
-  }
-}
 
 export function buildReducer(reducerMapping, reducers) {
   return reducers.map((r) => {
@@ -25,79 +15,6 @@ export function buildReducer(reducerMapping, reducers) {
     }
     return reducer(r?.args)
   })
-}
-
-function extractFeeds(adapter) {
-  const adapterId = adapter.id
-  const feeds = adapter.feeds.map((f) => {
-    return {
-      url: f.url,
-      headers: f.headers,
-      method: f.method,
-      reducers: f.reducers
-    }
-  })
-
-  return { [adapterId]: { name: adapter.name, decimals: adapter.decimals, feeds } }
-}
-
-function extractAggregators(aggregator) {
-  const aggregatorAddress = aggregator.address
-  return {
-    [aggregatorAddress]: {
-      id: aggregator.id,
-      address: aggregator.address,
-      name: aggregator.name,
-      active: aggregator.active,
-      fixedHeartbeatRate: aggregator.fixedHeartbeatRate,
-      randomHeartbeatRate: aggregator.randomHeartbeatRate,
-      threshold: aggregator.threshold,
-      absoluteThreshold: aggregator.absoluteThreshold,
-      adapterId: aggregator.adapterId
-    }
-  }
-}
-
-function validateAdapter(adapter): IAdapter {
-  // TODO extract properties from Interface
-  const requiredProperties = ['id', 'active', 'name', 'jobType', 'decimals', 'feeds']
-  // TODO show where is the error
-  const hasProperty = requiredProperties.map((p) =>
-    Object.prototype.hasOwnProperty.call(adapter, p)
-  )
-  const isValid = hasProperty.every((x) => x)
-
-  if (isValid) {
-    return adapter as IAdapter
-  } else {
-    throw new OraklError(OraklErrorCode.InvalidAdapter)
-  }
-}
-
-function validateAggregator(adapter): IAggregator {
-  // TODO extract properties from Interface
-  const requiredProperties = [
-    'id',
-    'address',
-    'active',
-    'name',
-    'fixedHeartbeatRate',
-    'randomHeartbeatRate',
-    'threshold',
-    'absoluteThreshold',
-    'adapterId'
-  ]
-  // TODO show where is the error
-  const hasProperty = requiredProperties.map((p) =>
-    Object.prototype.hasOwnProperty.call(adapter, p)
-  )
-  const isValid = hasProperty.every((x) => x)
-
-  if (isValid) {
-    return adapter as IAggregator
-  } else {
-    throw new OraklError(OraklErrorCode.InvalidAggregator)
-  }
 }
 
 export function uniform(a: number, b: number): number {
